pyproject.toml
[project]
name = "article-silicon-sampling-quebec"
version = "0.1.0"
description = "Code workspace for silicon sampling research on Quebec public opinion"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "polars>=1.0.0",
  "pandas>=2.0.0",
  "pyarrow>=14.0.0",
  "python-dotenv>=1.0.0",
  "litellm>=1.0.0",
  "sentence-transformers>=2.0.0",
  "scikit-learn>=1.3.0",
  "huggingface_hub>=0.20.0",
  "transformers>=4.40.0",
  "torch>=2.0.0",
  "tokenizers>=0.19.0",
]

[project.optional-dependencies]
finetune = [
  "peft>=0.10.0",
  "trl>=0.8.0",
  "bitsandbytes>=0.43.0",
  "accelerate>=0.29.0",
  "datasets>=2.18.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
